requirements.txt
numpy~=1.23.3
pyyaml~=6.0
pandas~=1.5.0
tqdm~=4.64.1
scipy~=1.9.1
wandb~=0.13.1
torchmetrics~=0.9.3
matplotlib~=3.6.0
scikit-learn~=1.1.2
opencv-python~=4.6.0.66
pillow~=9.2.0
dataclasses